requirements.txt
pandas 
numpy 
matplotlib
scipy
https://download.pytorch.org/whl/cu100/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
torchvision
glibc
tqdm
librosa
tensorboardX